requirements.txt
audiomath==1.17.1
evdev==1.6.0
pafy==0.5.5
pexpect==4.8.0
Pillow==9.3.0
praw==7.6.1
pyjokes==0.6.0
python_vlc==3.0.16120
python_weather==0.3.8
pyttsx3==2.90
SpeechRecognition==3.8.1
thefuzz==0.19.0
translate==3.6.1
ttkthemes==3.2.2
wikipedia==1.4.0
